src/test/java/org/anne/aoc2024/Day4Test.java
package org.anne.aoc2024;

import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day4Test {

    List<String> input = Arrays.asList(
            "MMMSXXMASM",
            "MSAMXMSMSA",
            "AMXSXMAAMM",
            "MSAMASMSMX",
            "XMASAMXAMM",
            "XXAMMXXAMA",
            "SMSMSASXSS",
            "SAXAMASAAA",
            "MAMMMXMMMM",
            "MXMXAXMASX"
    );

    @Test
    void testPart1() {
        assertEquals(18, Day4.part1(input));
    }

    @Test
    void testPart2() {
        assertEquals(9, Day4.part2(input));
    }
}
